try a fourth attempt to fix Auth test
